refactor: 코드리뷰 반영 및 라우팅 재정의
