+++ packages/mobile/src/page/Map/style.module.scss
@@ -30,7 +30,7 @@ $linkTooltipDistance: 17px;
   z-index: 100;
 }
 
-.place_wrap {
+.balloon {
   position: fixed;
   bottom: $footerHeight + $footerLinkDistance + $linkHeight + $linkTooltipDistance;
   right: 26px;
@@ -39,8 +39,28 @@ $linkTooltipDistance: 17px;
   background-color: $white;
   font-size: 12px;
   letter-spacing: -0.02em;
-  line-height: 17px;
+  line-height: 12px;
   color: $black-80;
+
+  .content {
+    display: flex;
+    align-items: center;
+
+    svg {
+      margin-right: 3px;
+    }
+  }
+}
+
+.balloon:after {
+  position: absolute;
+  top: 60px;
+  left: 70px;
+  border-top: 15px solid #fff;
+  border-left: 20px solid transparent;
+  border-right: 0px solid transparent;
+  border-bottom: 0px solid transparent;
+  content: "";
 }
 
 .close_button {
